POC for comment.
Make the cost check pluggable so that domain specific logic is separated out from the generic transpose optimizer code (which ideally knows as little as possible about ORT specifics).

Create a default, default ORT, and post-layout transform implementations of the cost checker.

If post-layout transform we aggressively push the layout transform nodes, and fall-back to the regular cost check for all others. This differs from the existing approach of disabling the cost check for all nodes, and is hopefully a better approach as it limits the 'push at all costs' to the NCHW/NHWC Transpose nodes.